requirements.txt
rich
colorama
requests
pillow